Perl Scripts
NCBI_search.pl was used to get the sequences, sequences was used to
filter those results by unique names, comparer to identify two lists and
show those that are found in both lists or those that are found only
once. Counter2 is used to count percentage identity in a list of amino
acids.